4 new files
balanced_data.csv contains a data frame with the same number of fraud to nonfraud samples.
gans.py cote to generate more fraud samples (still to be tested).
generator_model_100.h5  is the download gans model. ( this model can be imported in gans.py)
umapPOnFroud.py is a 2d representation of the nonfraud and fraud data. ( using the balanced data)